LayoutTests/sputnik/Conformance/12_Statement/12.11_The_switch_Statement/S12.11_A3_T3-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Parse error
S12.11_A3_T3

PASS Expected parsing failure

TEST COMPLETE
